Rename importOnMedia to importOnMediaQuery
