+++ packages/js-toolkit/helpers/index.ts
@@ -5,4 +5,4 @@ export { default as getInstanceFromElement } from './getInstanceFromElement.js';
 export { default as importOnInteraction } from './importOnInteraction.js';
 export { default as importWhenIdle } from './importWhenIdle.js';
 export { default as importWhenVisible } from './importWhenVisible.js';
-export { default as importOnMedia } from './importOnMedia.js';
+export { default as importOnMediaQuery } from './importOnMediaQuery.js';
